requirements.txt
google-cloud-firestore
google-cloud-storage
beautifulsoup4
pandas
lxml
requests
numpy
